Don't deparallelize when building Tamarin
This was in there to work around (probably) a bug in Stack, which led
to a race condition and an irritating error on compiles. We've
upgraded Stack versions since then and it seems that the bug is no
longer present, so remove the deparallelization so that Stack can work
with all cores.